+++ pkg/server/xecho/handler_test.go
@@ -16,16 +16,83 @@ package xecho
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
+	"github.com/douyu/jupiter/pkg/util/xerror"
 	"github.com/douyu/jupiter/proto/testproto/v1"
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
 )
+
+type impl struct {
+	testproto.UnimplementedGreeterServiceServer
+}
+
+func (*impl) SayHello(ctx context.Context, req *testproto.SayHelloRequest) (*testproto.SayHelloResponse, error) {
+	if req.Name != "bob" {
+		return &testproto.SayHelloResponse{
+			Error: uint32(xerror.InvalidArgument.GetEcode()),
+			Msg:   "invalid name",
+			Data:  &testproto.SayHelloResponse_Data{},
+		}, nil
+	}
+
+	return &testproto.SayHelloResponse{
+		Msg: "",
+		Data: &testproto.SayHelloResponse_Data{
+			Name: "hello bob",
+		},
+	}, nil
+}
+
+func echoJsonHandler(c echo.Context) error {
+	req := new(testproto.SayHelloRequest)
+	err := c.Bind(req)
+	if err != nil {
+		return err
+	}
+
+	if req.Name != "bob" {
+		return c.JSON(http.StatusOK, &testproto.SayHelloResponse{
+			Error: uint32(xerror.InvalidArgument.GetEcode()),
+			Msg:   "invalid name",
+			Data:  &testproto.SayHelloResponse_Data{},
+		})
+	}
+	return c.JSON(http.StatusOK, &testproto.SayHelloResponse{
+		Msg: "",
+		Data: &testproto.SayHelloResponse_Data{
+			Name: "hello bob",
+		},
+	})
+}
+
+func echoHandler(c echo.Context) error {
+	req := new(testproto.SayHelloRequest)
+	err := c.Bind(req)
+	if err != nil {
+		return err
+	}
+
+	if req.Name != "bob" {
+		return c.JSON(http.StatusOK, &testproto.SayHelloResponse{
+			Error: uint32(xerror.InvalidArgument.GetEcode()),
+			Msg:   "invalid name",
+			Data:  &testproto.SayHelloResponse_Data{},
+		})
+	}
+	return ProtoJSON(c, http.StatusOK, &testproto.SayHelloResponse{
+		Msg: "",
+		Data: &testproto.SayHelloResponse_Data{
+			Name: "hello bob",
+		},
+	})
+}
 
 func TestGRPCProxyWrapper(t *testing.T) {
 	type args struct {
